Refactoring of models.py and rag.py
