+++ backend/ai_engine/app/question_gen/utils/security_utils.py
@@ -1,0 +1,8 @@
+import re
+
+def hide_urls(text: str) -> str:
+    url_pattern = (
+        r"http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\\(\\),]|"
+        r"(?:%[0-9a-fA-F][0-9a-fA-F]))+"
+    )
+    return re.sub(url_pattern, "<url-hidden>", text)
